+++ tools/nightcafe-challenge-generator/src/index.ts
@@ -2,6 +2,7 @@
 export { ChallengeGenerator } from './ChallengeGenerator';
 export { ThemeMatcher } from './ThemeMatcher';
 export { ChallengeLibraryManager } from './ChallengeLibraryManager';
+export { NightCafeHistoryManager } from './NightCafeHistoryManager';
 export { OutputFormatter, PrettyPrintFormatter, MarkdownFormatter, JsonFormatter, OutputFormatterFactory } from './OutputFormatter';
 export { OutputManager } from './OutputManager';
 export { CheatSheetScraper, NightCafeScraper } from './Scrapers';